Add channel manager to support multiple sensors
Add the ChannelManager class to support multiple ANT channels. This
commit implements all of the necessarry ANT features for connecting to
multiple sensors, but does not alter any of the UI elements. The
required UI updates to support these changes are still to come.

The ChannelManager is responsible for opening new search channels when
necesary, so WhimChannel's automatic reopening when a search timeout
occurs has been removed.